feat: lockfile global & route
Publish lockfile info in RESTful. I'm using a global variable to deal with it. There's  quite probably a better way of dealing with it. But for now I'm quite happy with the result. The program is able to open without the league client, and the file watcher seems to be working just fine.