fix: don't join thread
Joining threads makes the program wait for the thread to finish. But in this case, this thread will never stop as long as the program is running, so it breaks other things in the app. For now, I'm rolling it back until I figure how to handle the errors in a new thread.